+++ src/routes/search/Search.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import withStyles from 'isomorphic-style-loader/lib/withStyles';
 import s from './Search.css';
-import subjectItem from '../../components/SubjectItem'
+import SubjectItem from '../../components/SubjectItem';
 
 class Search extends React.Component {
   constructor(props) {
@@ -22,9 +22,9 @@ class Search extends React.Component {
           <h2 hidden={this.props.searchValue == ''}>
             Searching for {this.props.searchValue}
           </h2>
-          <subjectItem />
-
-
+          <div hidden={this.props.searchValue == ''}>
+            <SubjectItem subjectTitle={this.props.searchValue}/>
+          </div>
           <div hidden={this.props.searchValue != ''} className={s.searchBarEmptyTitle}>
             Please enter something in the searchbar, or press backspace to go
             back to the news